+++ packages/std/src/mock.rs
@@ -6,7 +6,7 @@ use crate::coins::Coin;
 use crate::encoding::Binary;
 use crate::errors::{contract_err, StdResult, Utf8StringErr};
 use crate::query::{AllBalanceResponse, BalanceResponse, BankQuery, QueryRequest, WasmQuery};
-use crate::serde::to_vec;
+use crate::serde::to_binary;
 use crate::storage::MemoryStorage;
 use crate::traits::{Api, Extern, Querier, QuerierResult};
 use crate::types::{BlockInfo, CanonicalAddr, ContractInfo, Env, HumanAddr, MessageInfo};
@@ -21,7 +21,7 @@ pub fn mock_dependencies(
 ) -> Extern<MockStorage, MockApi, MockQuerier> {
     let contract_addr = HumanAddr::from(CONTRACT_ADDR);
     Extern {
-        storage: MockStorage::new(),
+        storage: MockStorage::default(),
         api: MockApi::new(canonical_length),
         querier: MockQuerier::new(&[(&contract_addr, contract_balance)]),
     }
@@ -34,7 +34,7 @@ pub fn mock_dependencies_with_balances(
     balances: &[(&HumanAddr, &[Coin])],
 ) -> Extern<MockStorage, MockApi, MockQuerier> {
     Extern {
-        storage: MockStorage::new(),
+        storage: MockStorage::default(),
         api: MockApi::new(canonical_length),
         querier: MockQuerier::new(balances),
     }
@@ -124,48 +124,45 @@ pub fn mock_env<T: Api, U: Into<HumanAddr>>(api: &T, sender: U, sent: &[Coin]) -
 
 /// MockQuerier holds an immutable table of bank balances
 /// TODO: also allow querying contracts
-#[derive(Clone)]
+#[derive(Clone, Default)]
 pub struct MockQuerier {
-    balances: HashMap<HumanAddr, Vec<Coin>>,
+    bank: BankQuerier,
+    #[cfg(feature = "staking")]
+    staking: staking::StakingQuerier,
 }
 
 impl MockQuerier {
+    #[cfg(not(feature = "staking"))]
     pub fn new(balances: &[(&HumanAddr, &[Coin])]) -> Self {
-        let mut map = HashMap::new();
-        for (addr, coins) in balances.iter() {
-            map.insert(HumanAddr::from(addr), coins.to_vec());
+        MockQuerier {
+            bank: BankQuerier::new(balances),
         }
-        MockQuerier { balances: map }
+    }
+
+    #[cfg(feature = "staking")]
+    pub fn new(balances: &[(&HumanAddr, &[Coin])]) -> Self {
+        MockQuerier {
+            bank: BankQuerier::new(balances),
+            staking: staking::StakingQuerier::default(),
+        }
+    }
+
+    #[cfg(feature = "staking")]
+    pub fn with_staking(
+        &mut self,
+        validators: &[crate::query::Validator],
+        delegations: &[crate::query::Delegation],
+    ) {
+        self.staking = staking::StakingQuerier::new(validators, delegations);
     }
 }
 
 impl Querier for MockQuerier {
     fn query(&self, request: &QueryRequest) -> QuerierResult {
         match request {
-            QueryRequest::Bank(BankQuery::Balance { address, denom }) => {
-                // proper error on not found, serialize result on found
-                let amount = self
-                    .balances
-                    .get(address)
-                    .and_then(|v| v.iter().find(|c| &c.denom == denom).map(|c| c.amount))
-                    .unwrap_or_default();
-                let bank_res = BalanceResponse {
-                    amount: Coin {
-                        amount,
-                        denom: denom.to_string(),
-                    },
-                };
-                let api_res = to_vec(&bank_res).map(Binary).map_err(|e| e.into());
-                Ok(api_res)
-            }
-            QueryRequest::Bank(BankQuery::AllBalances { address }) => {
-                // proper error on not found, serialize result on found
-                let bank_res = AllBalanceResponse {
-                    amount: self.balances.get(address).cloned().unwrap_or_default(),
-                };
-                let api_res = to_vec(&bank_res).map(Binary).map_err(|e| e.into());
-                Ok(api_res)
-            }
+            QueryRequest::Bank(bank_query) => self.bank.query(bank_query),
+            #[cfg(feature = "staking")]
+            QueryRequest::Staking(staking_query) => self.staking.query(staking_query),
             QueryRequest::Wasm(msg) => {
                 let addr = match msg {
                     WasmQuery::Smart { contract_addr, .. } => contract_addr,
@@ -178,11 +175,251 @@ impl Querier for MockQuerier {
     }
 }
 
+#[derive(Clone, Default)]
+struct BankQuerier {
+    balances: HashMap<HumanAddr, Vec<Coin>>,
+}
+
+impl BankQuerier {
+    fn new(balances: &[(&HumanAddr, &[Coin])]) -> Self {
+        let mut map = HashMap::new();
+        for (addr, coins) in balances.iter() {
+            map.insert(HumanAddr::from(addr), coins.to_vec());
+        }
+        BankQuerier { balances: map }
+    }
+
+    fn query(&self, request: &BankQuery) -> QuerierResult {
+        match request {
+            BankQuery::Balance { address, denom } => {
+                // proper error on not found, serialize result on found
+                let amount = self
+                    .balances
+                    .get(address)
+                    .and_then(|v| v.iter().find(|c| &c.denom == denom).map(|c| c.amount))
+                    .unwrap_or_default();
+                let bank_res = BalanceResponse {
+                    amount: Coin {
+                        amount,
+                        denom: denom.to_string(),
+                    },
+                };
+                Ok(to_binary(&bank_res).map_err(|e| e.into()))
+            }
+            BankQuery::AllBalances { address } => {
+                // proper error on not found, serialize result on found
+                let bank_res = AllBalanceResponse {
+                    amount: self.balances.get(address).cloned().unwrap_or_default(),
+                };
+                Ok(to_binary(&bank_res).map_err(|e| e.into()))
+            }
+        }
+    }
+}
+
+#[cfg(feature = "staking")]
+mod staking {
+    use crate::query::{
+        Delegation, DelegationsResponse, StakingQuery, Validator, ValidatorsResponse,
+    };
+    use crate::serde::to_binary;
+    use crate::traits::QuerierResult;
+
+    #[derive(Clone, Default)]
+    pub struct StakingQuerier {
+        validators: Vec<Validator>,
+        delegations: Vec<Delegation>,
+    }
+
+    impl StakingQuerier {
+        pub fn new(validators: &[Validator], delegations: &[Delegation]) -> Self {
+            StakingQuerier {
+                validators: validators.to_vec(),
+                delegations: delegations.to_vec(),
+            }
+        }
+
+        pub fn query(&self, request: &StakingQuery) -> QuerierResult {
+            match request {
+                StakingQuery::Validators {} => {
+                    let val_res = ValidatorsResponse {
+                        validators: self.validators.clone(),
+                    };
+                    Ok(to_binary(&val_res).map_err(|e| e.into()))
+                }
+                StakingQuery::Delegations {
+                    delegator,
+                    validator,
+                } => {
+                    let matches = |d: &&Delegation| {
+                        if let Some(val) = validator {
+                            if val != &d.validator {
+                                return false;
+                            }
+                        }
+                        &d.delegator == delegator
+                    };
+                    let delegations: Vec<_> =
+                        self.delegations.iter().filter(matches).cloned().collect();
+                    let val_res = DelegationsResponse { delegations };
+                    Ok(to_binary(&val_res).map_err(|e| e.into()))
+                }
+            }
+        }
+    }
+
+    #[cfg(test)]
+    mod test {
+        use super::*;
+
+        use crate::{coin, from_binary, HumanAddr};
+
+        #[test]
+        fn staking_querier_validators() {
+            let val1 = Validator {
+                address: HumanAddr::from("validator-one"),
+                commission: 1000,
+                max_commission: 3000,
+                max_change_rate: 1000,
+            };
+            let val2 = Validator {
+                address: HumanAddr::from("validator-two"),
+                commission: 1500,
+                max_commission: 4000,
+                max_change_rate: 500,
+            };
+
+            let staking = StakingQuerier::new(&[val1.clone(), val2.clone()], &[]);
+
+            // one match
+            let raw = staking
+                .query(&StakingQuery::Validators {})
+                .unwrap()
+                .unwrap();
+            let vals: ValidatorsResponse = from_binary(&raw).unwrap();
+            assert_eq!(vals.validators, vec![val1, val2]);
+        }
+
+        // gets delegators from query or panic
+        fn get_delegators(
+            staking: &StakingQuerier,
+            delegator: HumanAddr,
+            validator: Option<HumanAddr>,
+        ) -> Vec<Delegation> {
+            let raw = staking
+                .query(&StakingQuery::Delegations {
+                    delegator,
+                    validator,
+                })
+                .unwrap()
+                .unwrap();
+            let dels: DelegationsResponse = from_binary(&raw).unwrap();
+            dels.delegations
+        }
+
+        #[test]
+        fn staking_querier_delegations() {
+            let val1 = HumanAddr::from("validator-one");
+            let val2 = HumanAddr::from("validator-two");
+
+            let user_a = HumanAddr::from("investor");
+            let user_b = HumanAddr::from("speculator");
+            let user_c = HumanAddr::from("hodler");
+
+            // we need multiple validators per delegator, so the queries provide different results
+            let del1a = Delegation {
+                delegator: user_a.clone(),
+                validator: val1.clone(),
+                amount: coin(100, "stake"),
+                can_redelegate: true,
+                accumulated_rewards: coin(5, "stake"),
+            };
+            let del2a = Delegation {
+                delegator: user_a.clone(),
+                validator: val2.clone(),
+                amount: coin(500, "stake"),
+                can_redelegate: true,
+                accumulated_rewards: coin(20, "stake"),
+            };
+
+            // this is multiple times on the same validator
+            let del1b = Delegation {
+                delegator: user_b.clone(),
+                validator: val1.clone(),
+                amount: coin(500, "stake"),
+                can_redelegate: false,
+                accumulated_rewards: coin(0, "stake"),
+            };
+            let del1bb = Delegation {
+                delegator: user_b.clone(),
+                validator: val1.clone(),
+                amount: coin(700, "stake"),
+                can_redelegate: true,
+                accumulated_rewards: coin(70, "stake"),
+            };
+
+            // and another one on val2
+            let del2c = Delegation {
+                delegator: user_c.clone(),
+                validator: val2.clone(),
+                amount: coin(8888, "stake"),
+                can_redelegate: true,
+                accumulated_rewards: coin(900, "stake"),
+            };
+
+            let staking = StakingQuerier::new(
+                &[],
+                &[
+                    del1a.clone(),
+                    del1b.clone(),
+                    del1bb.clone(),
+                    del2a.clone(),
+                    del2c.clone(),
+                ],
+            );
+
+            // get all for user a
+            let dels = get_delegators(&staking, user_a.clone(), None);
+            assert_eq!(dels, vec![del1a.clone(), del2a.clone()]);
+
+            // get all for user b
+            let dels = get_delegators(&staking, user_b.clone(), None);
+            assert_eq!(dels, vec![del1b.clone(), del1bb.clone()]);
+
+            // get all for user c
+            let dels = get_delegators(&staking, user_c.clone(), None);
+            assert_eq!(dels, vec![del2c.clone()]);
+
+            // for user with no delegations...
+            let dels = get_delegators(&staking, HumanAddr::from("no one"), None);
+            assert_eq!(dels, vec![]);
+
+            // filter a by validator (1 and 1)
+            let dels = get_delegators(&staking, user_a.clone(), Some(val1.clone()));
+            assert_eq!(dels, vec![del1a.clone()]);
+            let dels = get_delegators(&staking, user_a.clone(), Some(val2.clone()));
+            assert_eq!(dels, vec![del2a.clone()]);
+
+            // filter b by validator (2 and 0)
+            let dels = get_delegators(&staking, user_b.clone(), Some(val1.clone()));
+            assert_eq!(dels, vec![del1b.clone(), del1bb.clone()]);
+            let dels = get_delegators(&staking, user_b.clone(), Some(val2.clone()));
+            assert_eq!(dels, vec![]);
+
+            // filter c by validator (0 and 1)
+            let dels = get_delegators(&staking, user_c.clone(), Some(val1.clone()));
+            assert_eq!(dels, vec![]);
+            let dels = get_delegators(&staking, user_c.clone(), Some(val2.clone()));
+            assert_eq!(dels, vec![del2c.clone()]);
+        }
+    }
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
 
-    use crate::coins;
+    use crate::{coin, coins, from_binary};
 
     #[test]
     fn mock_env_arguments() {
@@ -234,5 +471,79 @@ mod test {
         let api = MockApi::new(10);
         let human = HumanAddr("longer-than-10".to_string());
         let _ = api.canonical_address(&human).unwrap();
+    }
+
+    #[test]
+    fn bank_querier_all_balances() {
+        let addr = HumanAddr::from("foobar");
+        let balance = vec![coin(123, "ELF"), coin(777, "FLY")];
+        let bank = BankQuerier::new(&[(&addr, &balance)]);
+
+        // all
+        let all = bank
+            .query(&BankQuery::AllBalances {
+                address: addr.clone(),
+            })
+            .unwrap()
+            .unwrap();
+        let res: AllBalanceResponse = from_binary(&all).unwrap();
+        assert_eq!(&res.amount, &balance);
+    }
+
+    #[test]
+    fn bank_querier_one_balance() {
+        let addr = HumanAddr::from("foobar");
+        let balance = vec![coin(123, "ELF"), coin(777, "FLY")];
+        let bank = BankQuerier::new(&[(&addr, &balance)]);
+
+        // one match
+        let fly = bank
+            .query(&BankQuery::Balance {
+                address: addr.clone(),
+                denom: "FLY".to_string(),
+            })
+            .unwrap()
+            .unwrap();
+        let res: BalanceResponse = from_binary(&fly).unwrap();
+        assert_eq!(res.amount, coin(777, "FLY"));
+
+        // missing denom
+        let miss = bank
+            .query(&BankQuery::Balance {
+                address: addr.clone(),
+                denom: "MISS".to_string(),
+            })
+            .unwrap()
+            .unwrap();
+        let res: BalanceResponse = from_binary(&miss).unwrap();
+        assert_eq!(res.amount, coin(0, "MISS"));
+    }
+
+    #[test]
+    fn bank_querier_missing_account() {
+        let addr = HumanAddr::from("foobar");
+        let balance = vec![coin(123, "ELF"), coin(777, "FLY")];
+        let bank = BankQuerier::new(&[(&addr, &balance)]);
+
+        // all balances on empty account is empty vec
+        let all = bank
+            .query(&BankQuery::AllBalances {
+                address: HumanAddr::from("elsewhere"),
+            })
+            .unwrap()
+            .unwrap();
+        let res: AllBalanceResponse = from_binary(&all).unwrap();
+        assert_eq!(res.amount, vec![]);
+
+        // any denom on balances on empty account is empty coin
+        let miss = bank
+            .query(&BankQuery::Balance {
+                address: HumanAddr::from("elsewhere"),
+                denom: "ELF".to_string(),
+            })
+            .unwrap()
+            .unwrap();
+        let res: BalanceResponse = from_binary(&miss).unwrap();
+        assert_eq!(res.amount, coin(0, "ELF"));
     }
 }